Move all existing JS actions to actions/javascript
