+++ .github/actions/javascript/getMergeCommitForPullRequest/index.js
