+++ .github/actions/javascript/getReleaseBody/getReleaseBody.js
@@ -1,7 +1,7 @@
 const _ = require('underscore');
 const core = require('@actions/core');
-const ActionUtils = require('../../libs/ActionUtils');
-const GithubUtils = require('../../libs/GithubUtils');
+const ActionUtils = require('../../../libs/ActionUtils');
+const GithubUtils = require('../../../libs/GithubUtils');
 
 // Parse the stringified JSON array of PR numbers, and cast each from String -> Number
 const PRList = _.map(ActionUtils.getJSONInput('PR_LIST', {required: true}), Number);
