+++ .github/scripts/buildActions.sh
@@ -5,7 +5,7 @@
 
 # In order for this script to be safely run from anywhere, we cannot use the raw relative path '../actions'
 declare ACTIONS_DIR
-ACTIONS_DIR="$(dirname "$(dirname "$0")")/actions"
+ACTIONS_DIR="$(dirname "$(dirname "$0")")/actions/javascript"
 
 # List of paths to all JS files that implement our GH Actions
 declare -r GITHUB_ACTIONS=(
